backend/pyproject.toml
[project]
name = "interview-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi[all,standard]>=0.115.12",
    "google-ai-generativelanguage>=0.6.15",
    "google-api-core>=2.24.2",
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.25.5",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.2",
    "websockets>=15.0.1",
]
